feat: 프로젝트 카드에 발표시간이 나오게 구현
